refactor(step5): domain, controller 패키지 추가
