+++ src/main/java/step3/game/domain/RandomGenerator.java
@@ -1,4 +1,4 @@
-package step3.game;
+package step3.game.domain;
 
 import java.util.Random;
