+++ src/main/java/step3/game/domain/car/Car.java
@@ -1,4 +1,4 @@
-package step3.game.car;
+package step3.game.domain.car;
 
 public class Car {
