+++ src/main/java/step3/game/domain/car/CarName.java
@@ -1,4 +1,4 @@
-package step3.game.car;
+package step3.game.domain.car;
 
 import java.util.Objects;
