+++ src/main/java/step3/view/output/OutputView.java
@@ -1,6 +1,6 @@
 package step3.view.output;
 
-import step3.game.car.Car;
+import step3.game.domain.car.Car;
 
 import java.util.Set;
