+++ src/test/java/step3/game/RacingCarGameTest.java
@@ -4,7 +4,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import step3.game.car.CarName;
+import step3.game.domain.car.CarName;
+import step3.game.domain.PositiveNumber;
+import step3.game.domain.RacingCarGame;
 
 import java.util.HashSet;
 import java.util.Set;
